数值转化函数修改

+++ common/basePage.py
@@ -2003,7 +2003,7 @@ end
 
         # 转换数字部分
         try:
-            num = float(num)
+            num = int(num)
         except ValueError:
             raise ValueError(f"Invalid numeric string: '{num}'")
